Remove old configuration service, use the new one
